code/optimize/requirements.txt
# 优化算法可视化项目依赖

# 核心数值计算库
numpy>=1.21.0
scipy>=1.7.0

# 可视化库
matplotlib>=3.5.0

# 符号计算（用于自定义函数）
sympy>=1.9.0

# 配置文件处理
PyYAML>=6.0

# 可选依赖（用于高级功能）
# 如果需要更好的动画性能，可以安装：
# pillow>=8.0.0  # 用于GIF保存
# ffmpeg-python>=0.2.0  # 用于视频保存

# 开发依赖（可选）
# pytest>=6.0.0  # 测试框架
# pytest-cov>=2.0.0  # 测试覆盖率
# black>=21.0.0  # 代码格式化
# mypy>=0.910  # 类型检查
# sphinx>=4.0.0  # 文档生成